Created separate components for default and moblie headers
